split schemas up
